+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-### **(V.1.1.1 Changes) (1.17.0 Minecraft)**
+### **(V.1.2.0 Changes) (1.17.0 Minecraft)**
 
 ##### Pillars:
 Fixed pillar code to be more accurate and not be messed up by caves and ravines.
+
+##### Config:
+Split removeWorldBottomStructures into removeStructuresOffIslands and removeWorldBottomPieces.
+ removeWorldBottomPieces applies to nbt pieces of any kind (not just structures).
+ removeStructuresOffIslands now will do better (hopefully) to remove any structure that goes off the island.
 
 
 ### **(V.1.1.0 Changes) (1.17.0 Minecraft)**